Move ImageServerUtil to operations module to make it usable by camel application
